refactor(script): extract wallet session wrapper
